# expose json encoding/decoding of prefix