--- a/prefix_test.go
+++ b/prefix_test.go
@@ -1206,3 +1206,19 @@ func TestPrefix_availablePrefixes(t *testing.T) {
 		})
 	}
 }
+
+func TestJSON(t *testing.T) {
+	testWithBackends(t, func(t *testing.T, ipam *ipamer) {
+		prefix, err := ipam.NewPrefix("192.168.0.0/24")
+		require.Nil(t, err)
+		require.NotNil(t, prefix)
+
+		data, err := prefix.JSONEncode()
+		require.Nil(t, err)
+
+		newPrefix := &Prefix{}
+		err = newPrefix.JSONDecode(data)
+		require.Nil(t, err)
+		require.Equal(t, prefix, newPrefix)
+	})
+}