@majst01 2021-01-26
You are right, i will close this, can always be looked at later if required.